- Moved acceptance tests to separate files - Added `TEST.md` with tests information (WIP). - Added some documentation for acceptance test functions and structures.
